backend/requirements.txt
django
djangorestframework
django-cors-headers
dj-database-url
python-dotenv
